firewall: Use new setup mechanism
- Also reorganize module to separate out views.
